[fx] use a linked list for nodes
This makes it possible to define reasonable semantics for what happens
when a node in the list is deleted. In particular the iteration over nodes
will continue at the node that was after the deleted node _when it was deleted_.
If the new node is also deleted, we skip it and, continue to the node after it.
Eventually we either reach a node still in the list or we reach the end of the list.

ghstack-source-id: 3f177bb38547837ccc505dc5f943e9e29c20a180
Pull Request resolved: https://github.com/pytorch/pytorch/pull/45708