Remove lock from GraphTask::set_exception_without_signal.
In most cases the lock ordering was hold a lock in local autograd and
then hold a lock in DistAutogradContext.

In case of `set_exception_without_signal` the lock order was in reverse and as
a result we saw potential deadlock issues in our TSAN tests. To fix this, I
removed the lock and instead just used std::atomic exchange.

In addition to this, I fixed TestE2E to ensure that we use the appropriate
timeout.

TestE2EProcessGroup was flaky for these two reasons and now is fixed.

Differential Revision: [D24120962](https://our.internmc.facebook.com/intern/diff/D24120962/)

ghstack-source-id: 113592709
Pull Request resolved: https://github.com/pytorch/pytorch/pull/45867